+++ CHANGELOG.rst
@@ -3,12 +3,14 @@ Changelog
 
 Version 0.8-dev
 ---------------
-| **Date:** August 06, 2021
+| **Date:** November 10, 2021
 | **Release:** v0.8-dev
 | **Version of CHAOS:** CHAOS-7.8 (0708)
 
 Features
 ^^^^^^^^
+* Added :func:`chaosmagpy.coordinate_utils.sh_analysis`, which performs a
+  spherical harmonic expansion.
 
 Version 0.7.1
 -------------